Add files via upload.
These files were part of my final deliverable for my group project for my data management course. These files are all part of a larger image editing program that takes an image and depending on user input, filters it a certain way. The whole program is written in Python. I was Team Leader for this project because I am a Software Engineering major.